windows: quote the arguments manually
The `_spawn*()` family of functions do not quote the arguments,
therefore we have to do that. However, the `Path*()` functions are
limited to `MAX_PATH`. Let's borrow Git for Windows' quoting strategy to
ensure that everything is quoted correctly.

Technically, we could be a bit sloppy and not bother about releasing the
memory allocated by `quote()`, but I like to clean up after myself ;-)

The quoting code (which I authored) originates from Git for Windows'
`git-credential-helper-selector` and is hereby re-licensed under the MIT
license.